+++ client/src/common.rs
