+++ client/src/components/maker/maker_save_button.rs
@@ -5,7 +5,7 @@ use yew::{function_component, html, Callback, Properties, UseStateHandle};
 
 use crate::model::maker_page::MakerState;
 use crate::model::maker_page::PutMakerRequest;
-use crate::settings::api::backend_url;
+use crate::common::api::backend_url;
 
 #[derive(Properties, PartialEq)]
 pub struct SaveButtonProperty {
