+++ client/src/lib.rs
@@ -1,4 +1,4 @@
 mod components;
 mod model;
 mod pages;
-mod settings;
+mod common;
