+++ client/src/pages/item/item_edit.rs
@@ -3,8 +3,8 @@ use crate::components::common::text_box::TextBox;
 use crate::components::item::item_detail::ItemDetail;
 use crate::model::common::NameOptionIdPair;
 use crate::model::item_page::{GetItemInfoByTitleId, ItemState, TitleInfo, TitleState};
-use crate::settings::api::backend_url;
-use crate::settings::select::{announce_status_list, catalog_status_list, project_type_list};
+use crate::common::api::backend_url;
+use crate::common::select::{announce_status_list, catalog_status_list, project_type_list};
 use uuid::Uuid;
 use reqwasm::http::Request;
 use urlencoding::decode;
